# apps/web/src/api/generated/openapi.json
{
 "components": {
  "schemas": {
   "AiProfile": {
    "description": "One AI profile (keys surface only as ``keyConfigured``).",
    "properties": {
     "baseUrl": {
      "title": "Baseurl",
      "type": "string"
     },
     "createdAt": {
      "title": "Createdat",
      "type": "string"
     },
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "keyConfigured": {
      "title": "Keyconfigured",
      "type": "boolean"
     },
     "label": {
      "title": "Label",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "provider": {
      "const": "openai_compatible",
      "title": "Provider",
      "type": "string"
     },
     "updatedAt": {
      "title": "Updatedat",
      "type": "string"
     }
    },
    "required": [
     "id",
     "label",
     "provider",
     "baseUrl",
     "model",
     "enabled",
     "keyConfigured",
     "createdAt",
     "updatedAt"
    ],
    "title": "AiProfile",
    "type": "object"
   },
   "AiProfileCreate": {
    "description": "POST /api/v1/settings/ai/profiles body (metadata only, no key).",
    "properties": {
     "baseUrl": {
      "default": "",
      "title": "Baseurl",
      "type": "string"
     },
     "enabled": {
      "default": true,
      "title": "Enabled",
      "type": "boolean"
     },
     "label": {
      "minLength": 1,
      "title": "Label",
      "type": "string"
     },
     "model": {
      "default": "",
      "title": "Model",
      "type": "string"
     }
    },
    "required": [
     "label"
    ],
    "title": "AiProfileCreate",
    "type": "object"
   },
   "AiProfileUpdate": {
    "description": "PATCH /api/v1/settings/ai/profiles/{id} body (all optional).",
    "properties": {
     "baseUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseurl"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     }
    },
    "title": "AiProfileUpdate",
    "type": "object"
   },
   "AiPurposeStatus": {
    "description": "Effective, secret-free resolution of one AI purpose.",
    "properties": {
     "baseUrl": {
      "title": "Baseurl",
      "type": "string"
     },
     "configured": {
      "title": "Configured",
      "type": "boolean"
     },
     "keyConfigured": {
      "title": "Keyconfigured",
      "type": "boolean"
     },
     "keySource": {
      "enum": [
       "profile_secret",
       "default_secret",
       "env",
       "missing"
      ],
      "title": "Keysource",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "profileId": {
      "title": "Profileid",
      "type": "string"
     },
     "profileLabel": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Profilelabel"
     },
     "source": {
      "enum": [
       "default",
       "profile"
      ],
      "title": "Source",
      "type": "string"
     }
    },
    "required": [
     "profileId",
     "source",
     "baseUrl",
     "model",
     "keyConfigured",
     "keySource",
     "configured"
    ],
    "title": "AiPurposeStatus",
    "type": "object"
   },
   "AiPurposesUpdate": {
    "description": "PUT /api/v1/settings/ai/purposes body (each purpose optional).\n\nEvery target is either the built-in ``default`` resolution or an\nexisting profile id.",
    "properties": {
     "chat": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Chat"
     },
     "summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     },
     "translation": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Translation"
     }
    },
    "title": "AiPurposesUpdate",
    "type": "object"
   },
   "AiSettingsUpdate": {
    "additionalProperties": false,
    "description": "PUT /api/v1/settings/ai body — every field optional, validated.\n\nMissing fields keep their current value; blank baseUrl/model clear it.\nThere is no secret field on this model by design, and unknown fields\nare rejected (extra=\"forbid\") so an attempt to smuggle e.g. an apiKey\ninto the settings store fails loudly instead of being ignored.",
    "properties": {
     "baseUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseurl"
     },
     "libretranslateUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Libretranslateurl"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "provider": {
      "anyOf": [
       {
        "const": "openai_compatible",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Provider"
     },
     "summaryLanguage": {
      "anyOf": [
       {
        "enum": [
         "zh-CN",
         "en"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summarylanguage"
     },
     "translationEngine": {
      "anyOf": [
       {
        "enum": [
         "ai",
         "libretranslate",
         "browser"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Translationengine"
     },
     "translationLanguage": {
      "anyOf": [
       {
        "enum": [
         "zh-CN",
         "en"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Translationlanguage"
     }
    },
    "title": "AiSettingsUpdate",
    "type": "object"
   },
   "AiSettingsView": {
    "description": "GET/PUT /api/v1/settings/ai (key presence is booleans only).",
    "properties": {
     "baseUrl": {
      "title": "Baseurl",
      "type": "string"
     },
     "configured": {
      "title": "Configured",
      "type": "boolean"
     },
     "defaultKeyConfigured": {
      "title": "Defaultkeyconfigured",
      "type": "boolean"
     },
     "envKeyConfigured": {
      "title": "Envkeyconfigured",
      "type": "boolean"
     },
     "libretranslateKeyConfigured": {
      "title": "Libretranslatekeyconfigured",
      "type": "boolean"
     },
     "libretranslateUrl": {
      "title": "Libretranslateurl",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "provider": {
      "const": "openai_compatible",
      "title": "Provider",
      "type": "string"
     },
     "purposeStatus": {
      "additionalProperties": {
       "$ref": "#/components/schemas/AiPurposeStatus"
      },
      "title": "Purposestatus",
      "type": "object"
     },
     "purposes": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Purposes",
      "type": "object"
     },
     "summaryLanguage": {
      "enum": [
       "zh-CN",
       "en"
      ],
      "title": "Summarylanguage",
      "type": "string"
     },
     "translationEngine": {
      "enum": [
       "ai",
       "libretranslate",
       "browser"
      ],
      "title": "Translationengine",
      "type": "string"
     },
     "translationLanguage": {
      "enum": [
       "zh-CN",
       "en"
      ],
      "title": "Translationlanguage",
      "type": "string"
     }
    },
    "required": [
     "provider",
     "baseUrl",
     "model",
     "summaryLanguage",
     "translationLanguage",
     "translationEngine",
     "libretranslateUrl",
     "libretranslateKeyConfigured",
     "configured",
     "envKeyConfigured",
     "defaultKeyConfigured",
     "purposes",
     "purposeStatus"
    ],
    "title": "AiSettingsView",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "SecretValuePut": {
    "description": "Write-only secret body (shared by AI keys and RSSHub secrets).\n\n0021 hardening: bounded length (same bound as the RSSHub secret\nschema) and control-character rejection at INPUT time — a control\ncharacter in an API key is never legitimate, and storing one would\npoison env-file rendering / upstream calls until the value is\nreplaced.",
    "properties": {
     "value": {
      "maxLength": 10000,
      "minLength": 1,
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "value"
    ],
    "title": "SecretValuePut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "LumiRSS BFF",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/settings/ai": {
   "get": {
    "description": "Current AI settings + purpose resolution.\n\nNo response field ever contains an API key: ``configured`` /\n``envKeyConfigured`` / ``defaultKeyConfigured`` /\n``purposeStatus.*.keyConfigured`` are booleans only.",
    "operationId": "get_ai_settings_api_v1_settings_ai_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AiSettingsView"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Ai Settings"
   },
   "put": {
    "description": "Persist non-secret GLOBAL (default) AI settings.\n\nThere is deliberately no API-key field here: keys live in the\nSecretsStore via the dedicated write-only key endpoints.",
    "operationId": "put_ai_settings_api_v1_settings_ai_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AiSettingsUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AiSettingsView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Ai Settings"
   }
  },
  "/api/v1/settings/ai/key": {
   "delete": {
    "description": "Remove the browser-set default key (env fallback resumes).",
    "operationId": "delete_default_ai_key_api_v1_settings_ai_key_delete",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    },
    "summary": "Delete Default Ai Key"
   },
   "put": {
    "description": "Store the default (legacy) AI API key from the browser.\n\nWrite-only: the value is persisted server-side in the SecretsStore\nand can never be read back over the API. The env ``AI_API_KEY``\nremains as fallback when no browser key is set.",
    "operationId": "put_default_ai_key_api_v1_settings_ai_key_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SecretValuePut"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Default Ai Key"
   }
  },
  "/api/v1/settings/ai/profiles": {
   "get": {
    "description": "All AI profiles (metadata + ``keyConfigured`` flag, never keys).",
    "operationId": "get_ai_profiles_api_v1_settings_ai_profiles_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AiProfile"
         },
         "title": "Response Get Ai Profiles Api V1 Settings Ai Profiles Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Ai Profiles"
   },
   "post": {
    "description": "Create a profile. The API key is set separately (write-only).",
    "operationId": "create_ai_profile_api_v1_settings_ai_profiles_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AiProfileCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AiProfile"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Ai Profile"
   }
  },
  "/api/v1/settings/ai/profiles/{profile_id}": {
   "delete": {
    "description": "Delete a profile, its secret, and any purpose mapping to it.",
    "operationId": "delete_ai_profile_api_v1_settings_ai_profiles__profile_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "profile_id",
      "required": true,
      "schema": {
       "title": "Profile Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Ai Profile"
   },
   "patch": {
    "description": "Update profile metadata (label / baseUrl / model / enabled).",
    "operationId": "patch_ai_profile_api_v1_settings_ai_profiles__profile_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "profile_id",
      "required": true,
      "schema": {
       "title": "Profile Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AiProfileUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AiProfile"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Ai Profile"
   }
  },
  "/api/v1/settings/ai/profiles/{profile_id}/secret": {
   "delete": {
    "description": "Revoke one profile's API key.",
    "operationId": "delete_ai_profile_secret_api_v1_settings_ai_profiles__profile_id__secret_delete",
    "parameters": [
     {
      "in": "path",
      "name": "profile_id",
      "required": true,
      "schema": {
       "title": "Profile Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Ai Profile Secret"
   },
   "put": {
    "description": "Set/replace one profile's API key (write-only, never echoed).",
    "operationId": "put_ai_profile_secret_api_v1_settings_ai_profiles__profile_id__secret_put",
    "parameters": [
     {
      "in": "path",
      "name": "profile_id",
      "required": true,
      "schema": {
       "title": "Profile Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SecretValuePut"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Ai Profile Secret"
   }
  },
  "/api/v1/settings/ai/purposes": {
   "get": {
    "description": "Purpose → profile id (or ``default``) mapping.",
    "operationId": "get_ai_purposes_api_v1_settings_ai_purposes_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "string"
         },
         "title": "Response Get Ai Purposes Api V1 Settings Ai Purposes Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Ai Purposes"
   },
   "put": {
    "description": "Assign profiles to purposes. ``default`` selects the global\nsettings + default key resolution.",
    "operationId": "put_ai_purposes_api_v1_settings_ai_purposes_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AiPurposesUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "string"
         },
         "title": "Response Put Ai Purposes Api V1 Settings Ai Purposes Put",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Ai Purposes"
   }
  },
  "/api/v1/settings/translation/libretranslate-key": {
   "delete": {
    "description": "Remove the optional LibreTranslate API key.",
    "operationId": "delete_libretranslate_key_api_v1_settings_translation_libretranslate_key_delete",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    },
    "summary": "Delete Libretranslate Key"
   },
   "put": {
    "description": "Write-only LibreTranslate API key (optional; empty string clears).",
    "operationId": "put_libretranslate_key_api_v1_settings_translation_libretranslate_key_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SecretValuePut"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Libretranslate Key"
   }
  },
...
 }
}